spam mail detection using machine learning
The spam mail detection project utilizes machine learning techniques to develop an accurate and efficient system for identifying and filtering spam emails. By leveraging labeled datasets and employing feature extraction and machine learning algorithms, the project aims to provide users with an effective tool to combat the persistent problem of spam mail. The project's code and relevant resources will be made available on GitHub, enabling others to learn from and contribute to the ongoing efforts in spam detection researc